+++ Windows/Timer/StoryTimer.cpp
@@ -22,7 +22,7 @@ namespace Timer {
     {
 
 		std::chrono::milliseconds runtime = std::chrono::duration_cast<std::chrono::milliseconds> ( std::chrono::system_clock::now() - m_currentTime );
-        return runtime.count();
+        return (long) runtime.count();
     }
 
 }
